Don't use visit_stmt incorrectly
Summary: In two places we were using ast.statement.visit_stmt to loop through the top-level statements. That's true if the top-level statement is itself Statements, but if it's not and there is just one statement, you end up missing it. Switch to a dedicated top_level_statements function which does what these things wanted to do.

Reviewed By: stepancheg

Differential Revision: D42607765

fbshipit-source-id: c4551854d4f11991d578b1bfe71687042160a561