SmallMap -> Box<[(,)]> in CoercedSelector.entries
Summary:
We never use selector entries as map, so store them as `Box<[]>` instead of `SmallMap`.

This way we save a little memory: no need to store hashes (or also store an index is the map is large).

Small detail: why the limit to branch whether use quadratic or linear check is done when size is 32. I tried limit 7 or 16, them seem to perform slightly worse.

Reviewed By: krallin

Differential Revision: D42540797

fbshipit-source-id: c6c955bf6bab3d6ce855339210e40dd86fc6e697