Better startup timeout
Summary:
Follow-up to [post](https://fb.workplace.com/groups/buck2users/posts/24034878189467470).

Buck startup fails with:

```
Caused by:
    0: When establishing connection to buckd
    1: Failed to establish a connection to the Buck daemon or start a daemon within 30s
```

When daemon timeout, I'd like to know where it times out.

So idea is this.

`StartupTimeout` utility holds current deadline for startup.

`StartupTimeout::down` operation accepts a functions, which is:
* called with timeout
* passed timeout - 100ms

So when call stack is `A -> B -> C`, A is called with timeout 30s, B is called with timeout 29.9s, and C is called with timeout 29.8s.

So if `C` times out, before this diff we get an error at `A` and no context, and now we get error at `C` and the context from the stack.

Not all paths are handled properly yet, but this is a start.

Also code is ugly. I don't know how to make errors with timeouts better.

Reviewed By: krallin

Differential Revision: D42198887

fbshipit-source-id: 4d98d1dc848431fe64801c5b15981e0beb6ac385