Send file watcher stats to ingress
Summary:
Sending a whole `FileWatcherStats` so ingress can add more information around file watcher events and so we can clean up the code in the next stacked diffs.

There are 2 redundant information in `InvocationRecord`, `file_changes_since_last_build` and `branched_from_revision`, and 1 truncating block in `scribe.rs`, which are kept for now for the purpose of the phasing rollout and eventually are removed at the last diff of this stack.

Reviewed By: krallin

Differential Revision: D42514016

fbshipit-source-id: 105f62b56313e85afd9e00e888aa3c926f77bc7e