Log count of dropped scribe messages in snapshots
Summary:
After landing D42178265, it appears we now routinely drop lots of messages (mostly relating to action data). krallin theorizes this could be because we're exhausting the internal buffer.

Testing on this diff seems to corroborate this. Let's log the number of messages that we drop in the snapshots we take of scribe sink metrics so we can keep an eye on this (and figure out how large we need to make the buffer).

Reviewed By: krallin

Differential Revision: D42648117

fbshipit-source-id: 74153561b7ea866f94c5914704ea84e7dc2ad9c5