buck2: event log: allow accessing I/O stats when reading log files
Summary:
Useful to be able to estimate how far along we are in a log file, but also to
measure size of serialized vs. deserialized data.

Note that we take `ReaderStats` here since the point is to let the caller be
able to see where we're at while reading. In theory, we *could* expose those
stats via the stream itself (which ultimately has a reference to the counters
inside it), but that gets complicated fast given the amount of wrappers we have
here and this seems a lot easier.

Reviewed By: stepancheg

Differential Revision: D42638497

fbshipit-source-id: d0a0f8a2e71b35a00dfbc19125386c1814484d45