Add internal attrs to AttributeSpec in tests
Summary:
The change

{F852071994}

And the rest is fix of tests.

This change is needed because following diff D42662069 makes changes which assume `AttributeSpec` contains `visibility` attribute.

Also it is better when tests behavior is similar to real behavior.

Also, internal attributes should not probably be stored together with regular attributes, but let's change one thing at a time.

Reviewed By: ndmitchell

Differential Revision: D42661957

fbshipit-source-id: 73cc364b387bcda746d6d9ece2e970e1c6267b32