ArcStrBase
Summary:
Extract generic code from `ArcStr`.

Used in the following diff D42651348 to implement generic `ThinArcStr`.

This diff also changes the length from `usize` to `u32`. Does not matter much, although in some cases may result in more efficient enum variant packing. This is done for uniformity with `ThinArcStr` which limits the size to `u32` where it matters.

# How it works

`ArcStrBase` type is parameterized by `ArcStrProperties` trait which basically specifies where to put the length: in the struct, or in the heap.

For `ArcStr` length in stored in the struct, and `ThinArcStr` (following diff D42651348) stores it in the heap.

Reviewed By: JakobDegen

Differential Revision: D42651345

fbshipit-source-id: 663190496f678713dac172a66113488993598308