buck2: more_futures: introduce a new mechanism for cancelling futures
Summary:
There are a couple reasons for this.

- The first one is that I'd like to support "async cancellation" of futures
  (i.e. futures that don't exist immediately when they are told to cancel), and
  to do this I need to have
- The second one is to fix some gaps in the current implementation. The current
  implementation cancels futures by dropping them on a different thread (i.e.
  not on the task that is polling the future), which means that futures that
  are cancelled are not guaranteed to be released by the executor. It also
  relies on quite a bit of complicated logic and `unsafe` to work, though I
  don't have evidence of bugs arising from this.

This new implementation doesn't attempt to drop "remote" futures from the thread
that released the last reference. Instead, when a future needs to be cancelled,
we notify its waker that it should exit.

Note that this doesn't necessarily imply the caller *must* wait for the future
to actually exit (though that's the end goal of this stack).

The external API is more or less the same (the next diff updates `spawn()` to
use this and doesn't require changes in DICE, only to a test that observed
an implementation detail).

This diff also removes an atomic update (and replaces it with only a read) when
executing DICE futures, though that seems rather unlikely to matter in
practice.

Reviewed By: bobyangyf

Differential Revision: D42552208

fbshipit-source-id: 6878e044ce692f68666f1b82301b54f394ec613d