buck2: dice: remove the `on_cleanup` mechanism in favor of a drop guard
Summary:
I'd explained the rationale for this in D40741789 (https://github.com/facebookincubator/buck2/commit/605ad76389543f7647e3cabbf563bbd5f2b0cf00):

> In particular, I don't think you need to have separate implementations of the
> removal in the finished & cancel cases, and I'm not sure you need the
> "on_cancel" mechanism at all.
>
> To do that, I think you could instantiate a struct that just has a Drop impl
> that does the clean up in new_dice_task. The upsides of this (assuming you
> agree it works) are:
>
> - Only one codepath for all cases
> - Don't need to make extra copies of the key or the engine

bobyangyf had pointed out that we can't quite do that because futures could get
evicted from the currently_running map before they actually finish, so to fix
that we add an epoch on all requests (this is also generally useful for
debugging anyway since we can connect currently running map entries to running
futures).

The reason I'd like to do this is because I'd like to see if we can support
async cancellation of tasks in DICE, and to do that I need to rewrite some of
the cancellation bits, and it's easier to do so after decoupling this cleanup
mechanism entirely.

Reviewed By: bobyangyf

Differential Revision: D42543834

fbshipit-source-id: 9a309401d96a9f3fded526dde5308e8c6e46ae7b