buck2: more_futures: use CancellableFuture in spawn
Summary:
Like it says in the title. This updates the `spawn` module to use
`CancellableFuture`. Note that there is a small change needed in a test: the
point of the change is that cancellation is now async, so the callsite that
tests for cancellation now needs to expect it to happen in time instead of
instantly.

Reviewed By: bobyangyf

Differential Revision: D42552211

fbshipit-source-id: 3f0dcb27b39e39b4721433d0351d1bbde9946a15