expose loaded symbol names in ast
Summary:
`AstModule` previously only had a way to see what modules were being loaded,
but it's useful to see what symbols are being loaded from those modules too.

This is probably not the right place for this, but putting this up for comments.
Happy to move it somewhere else / come up with something else if this is undesirable.

Reviewed By: stepancheg

Differential Revision: D42616939

fbshipit-source-id: 1e29dbd92f57581344e67a59f80225d18ffc93cf