Another attempt to fix incorrect FileNameBuf::unchecked_new call
Summary:
The problem is: `SimpleDirEnty` `file_name` field (of type `FileNameBuf`) was created without checking filename is valid (no backslashes).

Previous trivial attempt D42723591 failed: we check for ignored files after we read directory. So even if directory is ignored, buck2 fails to query one of directories in the repository with backslash in path.

This is more complex attempt to fix the issue (quite verbose, but more correct).

In addition to `SimpleDirEntry` this diff introduces `RawDirEntry` which stores filename as `CompactString` instead of `FileNameBuf`.

So `ReadDirOutput` fields have different types now:

```
pub struct ReadDirOutput {
    pub included: Arc<[SimpleDirEntry]>,
    pub ignored: Arc<[RawDirEntry]>,
}
```

This is what is returned from `IoProvider::read_dir`, and this is what is stored in ignored paths.

# Update

In previous attempt (in this diff) I quietly ignored bad paths. Thomas suggested it should be a warning so users could add them to ignores.

It is not easy to make a warning from inside `DiceFileOpsDelegate` (because `buck2_common` does not depend on `buck2_events`), so I did different thing instead:
* check ignored paths using `MaybeIgnoredCellRelativePath` (previous diff D42767965 (https://github.com/facebookincubator/buck2/commit/91d7643accbaa88adb9f450311370184010bad87))
* keep returning error on bad file names

Reviewed By: krallin

Differential Revision: D42732053

fbshipit-source-id: c59f907d0bbf0138e61d69dfb0af037dc325c555