Add superconsole displays for concurrent command
Summary:
When concurrent commands are run, the command that is waiting will display in the superconsole that it is waiting for XXX command to finish running. When the waiting command is done waiting, then this message will go away and the critical section logging will commence.

Note that the soft_errors are still being emitted, and they show the actual command and the full trace ID (since the span width may be too small to fit the full log).

Reviewed By: krallin

Differential Revision: D42748372

fbshipit-source-id: 061a9e3cf49285963a4c916dd076107555df305c