ArcS<PackageRelativePath> in package listing and in CoercedPathDirectory
Summary:
Before this diff we store two copies of paths: in package listing and in attributes.

This diff replaces `Box<PackageRelativePath>` with `ArcS<PackageRelativePath>` and copies are shared now.

Previous version of the diff introduced backward incompatibility for query. This is no longer the case.

Also previous version of the diff had a type

```
struct PackageRelativePath(Arc<str>);
```

This version uses more generic `ArcS` type.

Reviewed By: bobyangyf

Differential Revision: D41251486

fbshipit-source-id: 0fa2f3ae630277d1022117df0e2e414d417a0c66