buck2: data: box large variants
Summary:
Our BuckEvents are positively huge (in large part due to the InvocationRecord),
and that means that:

- Allocating memory to hold them is costly
- Holding them in memory is costly.
- Moving them in the various layers produce them is CPU intensive.

We can fix all this with boxing large variants and also boxing the things we
send over the wire.

Test Plan

Before:

```
[torozco@devbig053]~/fbsource2 % /bin/time -v buck2 debug log-perf ~/large.pb.zst | tail -n 1
  User time (seconds): 20.33
  System time (seconds): 0.29
  ...

5802940	159021997	1010038791	12214183595
```

After:

```
[torozco@devbig053]~/fbsource2 % /bin/time -v ~/buck2.boxed debug log-perf ~/large.pb.zst | tail -n 1
  User time (seconds): 12.11
  System time (seconds): 0.28
  ...

5802940	159021997	1010038791	3257346203
```

It took 60% of the time and allocated 25% of the memory to deserialize the
exact same events.

freerealestate

Reviewed By: scottcao

Differential Revision: D42781412

fbshipit-source-id: c33b7ab86db03d88ff126f6374a01e52b4c3a164