Add repository_aliases feature
Summary:
As suggested in D42757735, have a section of the config that specifies aliases for the repositories only, not canonical names. That ensures we have no dependence on the order within a config, but still get the real alias we want.

We should deprecate duplicate aliases not passed via this mechanism one day, but perhaps when Buck1 dies.

Reviewed By: stepancheg

Differential Revision: D42771110

fbshipit-source-id: be415e4c6df30519edd499c46a7e750844c6df3f