Refactor traits in per-package interner
Summary:
Doing this makes the next diff D42721910 easier.

This diff gets rid of the Internable trait and instead replaces it with an Equiv trait that just has an equivalent function. We don't need the convert function from Internable because we also always From/Into for the actual implementation, so just make `intern` function directly require `Into<T>`.
Also, added a blanket implementation for Equiv<T> for T.

Reviewed By: stepancheg

Differential Revision: D42697026

fbshipit-source-id: 7ecb7b895728fac093bd4c909e76ce39d7fe73ee