Add additional example of `AnyLifetime` unsoundness
Summary:
The original test case given to show that this was unsound used `downcast_mut`. That lead me to think that the fundamental reason for this unsoundness was variance related, and that this could be fixed by enforcing some suitable invariance in `downcast_mut`. This diff adds a test case to show that that's not the case. The "real bug" here is instead that the implicit contract associated with the lifetime on the trait seems to not extend past the `: 'a` bound. However, more aggressive assumptions are needed to justify `downcast_ref` and especially `downcast_mut`.

This particular case is fixable by adding a `T: 'static` bound to the `&'a T: AnyLifetime<'a>` impl. However, I have no reason to believe that that's enough, and so won't change any code for now.

I have some idea for how this API might be fixable, but I need to think about it more and am not so sure it'll work.

Reviewed By: ndmitchell, stepancheg

Differential Revision: D42869701

fbshipit-source-id: 674316671903dcb8fc77e5cd735ede52526c27f6