buck2: concurrency: make DiceCommandDataProvider reusable
Summary:
This makes the next diff easier. We don't appear to gain much from holding onto the transaction across iterations, not to mention that
this case should be somewhat unusual to begin with.

It also seems a little easier to reason about commands that queue up
as if they had simply arrived later.

To enable that, make it possible to reuse the DiceCommandDataProvider so we can restart transactions.

Reviewed By: bobyangyf

Differential Revision: D42534322

fbshipit-source-id: 93ce80fb9d4ad97033589a25a1fcac2ed99fcdff