buck2: concurrency: wait for DICE to go idle before reusing
Summary:
The goal of this module is to ensure nothing is running on DICE in two
incompatible transactions, but it doesn't quite do that.

Instead, it ensures two *callers* don't exist on incompatible transactions,
but that overlooks the fact that computations can continue to exist in DICE
after a caller is gone (until they are cancelled, which may take some time if
they are in any critical sections)

This diff fixes it by placing the active dice a temporary quarantine state
after all callers have disconnected.

We could go a little bit more detailed here and track individual versions,
but since we currently only allow one executing version, it seems just as
easy to simply expect all transactions to finish.

Reviewed By: bobyangyf

Differential Revision: D42534317

fbshipit-source-id: 3971dabed686ba6b2cd84bb4ed62333949e89da7