buck2: add a test for config hashes not changing
Summary:
Like it says in the title. We decided that we don't want those to change *by
accident* anymore because it does not play nice with fast rollbacks, gradual
rollouts, and VB-time perf testing.

So, let's encode the hash in a test. That doesn't mean we can't change it ever,
but we can't change it without meaning to.

Reviewed By: IanChilds

Differential Revision: D42990148

fbshipit-source-id: 84482a595212dd6d73a279f57c1af9074a0cbdd9