buck2: fix critical sections ... again
Summary:
I tried to fix this in D42920743 (https://github.com/facebookincubator/buck2/commit/bfe1c1eb91d475ea2982ff8ddcc5a9bb695ae44c), but I neglected a separate way in which
this can fail....

To recap, in the deferred materializer, we produce a future using
`blocking_io`, but then we never await it. We then stuff it in a `Shared` and
hand it off to callers to await.

This means that multiple callers may await this and they might all be
cancellable. To fix this, D42920743 (https://github.com/facebookincubator/buck2/commit/bfe1c1eb91d475ea2982ff8ddcc5a9bb695ae44c) updated the implementation so we don't
error out if a critical section is exited by a cancellable caller.

However, that doesn't solve the case where the materializer produces the future
but never awaits it (sometimes it does, sometimes it doesn't, but in case I'd
see when writing D42920743 (https://github.com/facebookincubator/buck2/commit/bfe1c1eb91d475ea2982ff8ddcc5a9bb695ae44c), it did), and then 2 different tasks both await
it, with one entering the critical section and the other one exiting.

To fix this, let's capture the calling future synchronously when calling
`critical_section`.

Ultimately, this is all a mess, to solve it we need to stop making this
implicit using thread locals.

Reviewed By: IanChilds

Differential Revision: D42996038

fbshipit-source-id: cb9c6ce4061e7ce566835e74d5eb66194fb2b3c9